refactor: move UI folder to components for clarity
