+++ src/components/footer/NewsLetter.tsx
@@ -5,8 +5,8 @@ import { FormValues } from '@/types';
 
 import { emailIcon } from '@/assets';
 
-import Button from '@/ui/Button/Button';
-import Input from '@/ui/Input/Input';
+import Button from '@/components/ui/Button/Button';
+import Input from '@/components/ui/Input/Input';
 
 const NewsLetter = () => {
 	const { register } = useForm<FormValues>();
